% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cantons}
\alias{cantons}
\title{Swiss cantonal boundaries}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 50 rows and 1 columns.
}
\source{
\url{https://www.swisstopo.admin.ch/de/geodata/landscape/boundaries3d.html#technische_details}
}
\usage{
cantons
}
\description{
Shapefile downloaded from swisstopo, read in the script data-raw/01_get_data.R
}
\keyword{datasets}
