% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capstonePrototype-package.R
\docType{package}
\name{capstonePrototype-package}
\alias{capstonePrototype}
\alias{capstonePrototype-package}
\title{capstonePrototype: Shiny app with energy consumption of the cities Zurich, Basel and Winterthur}
\description{
This app showcases the up-to-date power consumption for the three Swiss cities for which the data is openly available.
}
\author{
\strong{Maintainer}: Sara Hänzi \email{first.last@example.com}

}
\keyword{internal}
