% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_city_info.R
\name{city_info_server}
\alias{city_info_server}
\title{city_info Server}
\usage{
city_info_server(id, cantons, cities, bs_colors)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
city_info Server
}
\keyword{internal}
