% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dataviz.R
\name{datavizUI}
\alias{datavizUI}
\title{dataviz UI}
\usage{
datavizUI(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
UI of the dataviz module,
once for each city
}
\keyword{internal}
