% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_wrangle.R
\name{deal_with_ts_zh}
\alias{deal_with_ts_zh}
\title{deal_with_ts_zh}
\usage{
deal_with_ts_zh(zh)
}
\arguments{
\item{zh}{tibble with zurich data, expect timestamp column (as character)}
}
\value{
zh tibble with timestamp converted to CET
}
\description{
Convert Timestamp from Zurich data, which is given in local time,
to a timestamp with Central European Time
}
\keyword{internal}
