% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_data.R
\name{get_bs_data}
\alias{get_bs_data}
\title{get_bs_data}
\usage{
get_bs_data()
}
\value{
bs tibble
}
\description{
function to download the entire bs data as one csv (takes a while)
}
\keyword{internal}
