% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data_wrangle.R
\name{get_clean_data}
\alias{get_clean_data}
\title{get_clean_data}
\usage{
get_clean_data(df)
}
\arguments{
\item{df}{tibble with timestamp}
}
\value{
clean time series (can be converted to tsibble without error)
}
\description{
function to check integrity of time series: checks for and removes
duplicates, fills gaps
}
\keyword{internal}
