% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_data.R
\name{get_csv_from_link}
\alias{get_csv_from_link}
\title{get_csv_from_link}
\usage{
get_csv_from_link(my_url)
}
\arguments{
\item{my_url}{Parameter specifying the csv file's url.}
}
\value{
returns the data from the csv as a tibble
}
\description{
function to read a csv function from a file/link
}
\keyword{internal}
