% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_data.R
\name{get_data_one_year_bs}
\alias{get_data_one_year_bs}
\title{get_data_one_year_bs}
\usage{
get_data_one_year_bs(year)
}
\arguments{
\item{year}{YYYY to get the data for}
}
\value{
tibble
}
\description{
get_data_one_year_bs
}
\keyword{internal}
