% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_data.R
\name{get_winti_data_up_to_date}
\alias{get_winti_data_up_to_date}
\title{get_winti_data_up_to_date}
\usage{
get_winti_data_up_to_date()
}
\value{
wi_current tibble
}
\description{
function to download and pre-process the latest csv file for
Winterthur
}
\keyword{internal}
