% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_get_data.R
\name{get_zh_data_up_to_date}
\alias{get_zh_data_up_to_date}
\title{get_zh_data_up_to_date}
\usage{
get_zh_data_up_to_date()
}
\value{
zh_current tibble
}
\description{
function to download and pre-process the latest csv file for
Zurich
}
\keyword{internal}
