% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_dataviz.R
\name{heatmap_tod_date}
\alias{heatmap_tod_date}
\title{heatmap_tod_date}
\usage{
heatmap_tod_date(data)
}
\arguments{
\item{df}{tibble with timestamp_hours_only, date, gross_energy_kwh columns}
}
\value{
ggplot
}
\description{
heatmeap with time of day on x and date on y axis, energy as colour
A utility for the dataviiz module
}
\keyword{internal}
