% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_overview.R
\name{overview_server}
\alias{overview_server}
\title{overview Server}
\usage{
overview_server(
  id,
  data_for_plots_week,
  data_for_plots_month,
  data_for_plots_year,
  data_for_plots_year_cumulative,
  bs_colors
)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
overview Server
}
\keyword{internal}
