% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_overview.R
\name{overviewUI}
\alias{overviewUI}
\title{overview UI}
\usage{
overviewUI(id)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
UI of the overview module, which is shown on the first tab,
once for each city
}
\keyword{internal}
