% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_city_info.R
\name{plot_cities}
\alias{plot_cities}
\title{plot_cities}
\usage{
plot_cities(cantons, cities, bs_colors)
}
\arguments{
\item{cantons}{sf object with cantonal boundaries}

\item{cities}{small tibble with Winterthur, Zurich and Basel coordinates}

\item{bs_colors}{named hex color vector with "info", "danger" and "success" colors}
}
\value{
a ggplot object
}
\description{
plot_cities
}
\keyword{internal}
