% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_dataviz.R
\name{plot_daily_per_year}
\alias{plot_daily_per_year}
\title{plot_daily_per_year}
\usage{
plot_daily_per_year(data)
}
\arguments{
\item{data}{a tibble with timestamp, year, yday and gross_energy_kwh columns}
}
\value{
ggplot
}
\description{
Function to plot daily energy use values over days of the year
with one color per year, a utility for the dataviz module
}
\details{
drops days with fewer than 95 entries, calculates rolling average
}
\keyword{internal}
