% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_dataviz.R
\name{plot_monthly_per_year}
\alias{plot_monthly_per_year}
\title{plot_monthly_per_year}
\usage{
plot_monthly_per_year(data)
}
\arguments{
\item{data}{a tibble with timestamp, year, month and gross_energy_kwh columns}
}
\value{
ggplot
}
\description{
Function to plot monthly energy use values over months of the year
with one color per year, a utility for the dataviz module
}
\details{
drops months with fewer than 28*96 entries
}
\keyword{internal}
