% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_predictions.R
\name{plot_prediction}
\alias{plot_prediction}
\title{plot_prediction}
\usage{
plot_prediction(data, bs_colors)
}
\arguments{
\item{data}{fabletools forecast from predict_2_weeks}

\item{bs_colors}{named color vector with "secondary" color}
}
\value{
ggplot
}
\description{
plot_prediction
}
\keyword{internal}
