% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_predictions.R
\name{plot_prediction_and_actual}
\alias{plot_prediction_and_actual}
\title{plot_prediction_and_actual}
\usage{
plot_prediction_and_actual(data_predicted, data_actual, bs_colors)
}
\arguments{
\item{data_predicted}{fabletools forecast from predict_2_weeks}

\item{data_actual}{tsibble on which the forecast is based}

\item{bs_colors}{named color vector with "secondary" color}
}
\value{
ggplot
}
\description{
plot_prediction_and_actual
}
\keyword{internal}
