% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_overview_week.R
\name{plot_week_reference}
\alias{plot_week_reference}
\title{plot_week_reference}
\usage{
plot_week_reference(data_ref, data_current, bs_colors)
}
\arguments{
\item{data_ref}{data_ref output from prepare_data_for_weekly_plot function}

\item{data_current}{data_current output from prepare_data_for_weekly_plot function}

\item{bs_colors}{named hex color vector with "light", "secondary" and "success" colors}
}
\value{
ggplot
}
\description{
function to plot the current data vs. a reference range for a week.
A utility function for the overview module.
}
\keyword{internal}
