% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_dataviz.R
\name{plot_weekly_per_year}
\alias{plot_weekly_per_year}
\title{plot_weekly_per_year}
\usage{
plot_weekly_per_year(data)
}
\arguments{
\item{data}{a tibble with timestamp, isoyear, week and gross_energy_kwh columns}
}
\value{
ggplot
}
\description{
Function to plot weekly energy use values over weeks of the year
with one color per year, a utility for the dataviz module
}
\details{
drops weeks with fewer than 7*94 entries
}
\keyword{internal}
