% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_zh_details.R
\name{plot_zh_details_last_week}
\alias{plot_zh_details_last_week}
\title{plot_zh_details_last_week}
\usage{
plot_zh_details_last_week(zh_details_week, bs_colors)
}
\arguments{
\item{zh_details_week}{tibble prepared by prepare_zh_details_last_week}

\item{bs_colors}{named color vector with success and primary colors}
}
\value{
ggplot
}
\description{
function to plot the last week for the household/industry
as an area plot. a utility for the zh_details module
}
\keyword{internal}
