% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_zh_details.R
\name{plot_zh_details_last_year}
\alias{plot_zh_details_last_year}
\title{plot_zh_details_last_year}
\usage{
plot_zh_details_last_year(zh_details_last_year, bs_colors)
}
\arguments{
\item{zh_details_last_year}{tibble prepared by prepare_zh_details_last_year}

\item{bs_colors}{named color vector with success and primary colors}
}
\value{
ggplot
}
\description{
function to plot last year's monthly values for the household/industry
as an area plot. a utility for the zh_details module
}
\keyword{internal}
