% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_zh_details.R
\name{plot_zh_details_years}
\alias{plot_zh_details_years}
\title{plot_zh_details_years}
\usage{
plot_zh_details_years(zh_details_yearly, bs_colors)
}
\arguments{
\item{zh_details_yearly}{tibble prepared by prepare_zh_details_years}

\item{bs_colors}{named color vector with success and primary colors}
}
\value{
ggplot
}
\description{
function to plot one value per year for the household/industry
as an area plot. a utility for the zh_details module
}
\keyword{internal}
