% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_predictions.R
\name{predict_2_weeks}
\alias{predict_2_weeks}
\title{predict_2_weeks}
\usage{
predict_2_weeks(data)
}
\arguments{
\item{data}{tsibble with gross_energy_kwh and timestamp as index}
}
\value{
a fabletools forecast for the next two weeks
}
\description{
function to predict the next two weeks' energy used based on
the past use, simple seasonal model with SNAIVE
}
\keyword{internal}
