% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_overview_year.R
\name{prepare_data_for_yearly_plot}
\alias{prepare_data_for_yearly_plot}
\title{prepare_data_for_yearly_plot}
\usage{
prepare_data_for_yearly_plot(data, date_today, n_ref = 5)
}
\arguments{
\item{data}{tibble wi/bs/zh with year, month, sum_gross_energy columns}

\item{date_today}{today's date, e.g. with lubridate::today()}

\item{n_ref}{default 5, number of years to go back as reference (if 5,
previous 4 will be the reference)}
}
\value{
list of data_ref and data_current
}
\description{
function to prepare reference data and current data for
yearly plot. a utility function in the overview module.
}
\details{
goes back in time one month at a time from date_today until data is
present in the data tibble. Uses recursion for this.
}
\keyword{internal}
