% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_zh_details.R
\name{prepare_zh_details_last_week}
\alias{prepare_zh_details_last_week}
\title{prepare_zh_details_last_week}
\usage{
prepare_zh_details_last_week(zh_details)
}
\arguments{
\item{zh_details}{tibble with timestamp, value_n5, valuene7 columns}
}
\value{
tibble with values over one week values by ne (Netzebene)
}
\description{
wrangle zh_details data into values for last week in long format
}
\keyword{internal}
