% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_zh_details.R
\name{prepare_zh_details_last_year}
\alias{prepare_zh_details_last_year}
\title{prepare_zh_details_last_year}
\usage{
prepare_zh_details_last_year(zh_details)
}
\arguments{
\item{zh_details}{tibble with timestamp, value_n5, valuene7 columns}
}
\value{
tibble with monthly values for last year by ne (Netzebene)
}
\description{
wrangle zh_details data into values for last year in long format
}
\keyword{internal}
