% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_module_zh_details.R
\name{prepare_zh_details_years}
\alias{prepare_zh_details_years}
\title{prepare_zh_details_years}
\usage{
prepare_zh_details_years(zh_details)
}
\arguments{
\item{zh_details}{tibble with timestamp, value_n5, valuene7 columns}
}
\value{
tibble yearly values by ne (Netzebene)
}
\description{
wrangle zh_details data into values per year in long format
}
\keyword{internal}
