% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zh}
\alias{zh}
\title{Zurich power usage data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 144960 rows and 14 columns.
}
\source{
\url{https://data.stadt-zuerich.ch/dataset/ewz_bruttolastgang_stadt_zuerich/download/2019_ewz_bruttolastgang.csv}

\url{https://data.stadt-zuerich.ch/dataset/ewz_bruttolastgang_stadt_zuerich/download/2020_ewz_bruttolastgang.csv}

\url{https://data.stadt-zuerich.ch/dataset/ewz_bruttolastgang_stadt_zuerich/download/2021_ewz_bruttolastgang.csv}

\url{https://data.stadt-zuerich.ch/dataset/ewz_bruttolastgang_stadt_zuerich/download/2022_ewz_bruttolastgang.csv}
}
\usage{
zh
}
\description{
Open government data with a power consumption data point every 15min for the city of Winterthur
}
\keyword{datasets}
