% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{zh_details}
\alias{zh_details}
\title{Zurich power usage data split according to voltage supply level}
\format{
\subsection{\code{zh_details}}{

A data frame, growing every day, the main three columns (all the others are derived):
\describe{
\item{timestamp}{local time}
\item{value_ne5}{energy consumption in Netzebene 5, industry and users with higher voltage level supply, in kWh}
\item{value_ne7}{energy consumption in Netzebene 7, households and other users with low voltage supply level, in kWh}
}
}
}
\source{
\url{https://data.stadt-zuerich.ch/dataset/ewz_stromabgabe_netzebenen_stadt_zuerich}
}
\usage{
zh_details
}
\description{
Open government data with a power consumption data point every 15min for the city of Zurich but
split according to voltage level ("Netzebene")
}
\details{
Downloaded and pre-processed with the script data-raw/02_get_updated_data
}
\keyword{datasets}
