% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_zh_details.R
\name{zh_details_server}
\alias{zh_details_server}
\title{zh_details Server}
\usage{
zh_details_server(id, bs_colors)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{bs_colors}{named hex color vector with "success" and "primary" color}
}
\description{
zh_details Server
}
\keyword{internal}
