% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dataviz.R
\name{dataviz_server}
\alias{dataviz_server}
\title{dataviz Server}
\usage{
dataviz_server(id, data)
}
\arguments{
\item{id}{Unique id for module instance.}
}
\description{
dataviz Server
}
\keyword{internal}
