% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dataviz.R
\name{datavizUI}
\alias{datavizUI}
\title{dataviz UI}
\usage{
datavizUI(id, choices_years)
}
\arguments{
\item{id}{Unique id for module instance.}

\item{choices_years}{vector of possible years to be chosen from the data}
}
\description{
UI of the dataviz module,
once for each city
}
\keyword{internal}
