% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wi}
\alias{wi}
\title{Winterthur power usage data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 355200 rows and 14 columns.
}
\source{
\url{https://www.web.statistik.zh.ch/ogd/daten/ressourcen/KTZH_00001863_00003561.csv}

\url{https://www.web.statistik.zh.ch/ogd/daten/ressourcen/KTZH_00001863_00003563.csv}

\url{https://www.web.statistik.zh.ch/ogd/daten/ressourcen/KTZH_00001863_00003564.csv}
}
\usage{
wi
}
\description{
Open government data with a power consumption data point every 15min for the city of Winterthur
}
\keyword{datasets}
